Add documentation and remove references to int64

+++ goterpreter/main.go
@@ -16,11 +16,11 @@ var (
 )
 
 // error handling.
-func reportError(line int64, msg string) {
+func reportError(line int, msg string) {
 	report(line, "", msg)
 }
 
-func report(line int64, where string, msg string) {
+func report(line int, where string, msg string) {
 	fmt.Fprintf(os.Stderr, "[line %d] Error %s: %s\n", line, where, msg)
 
 	hadError = true
